implement a tabbed layout in the webview UI
