+++ webview-ui/src/components/AssetLineage.vue
@@ -1,0 +1,14 @@
+<template>
+
+    <div class="flex flex-col p-4 space-y-4">
+      <div class="flex flex-col space-y-3">
+        <div class="flex flex-wrap gap-y-4">
+            <p>
+                This is supposed to be the asset lineage component
+            </p>
+        </div>
+    </div>
+    </div>
+  </template>
+  
+
